+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/iOSTestApp-Bridging-Header.h
@@ -4,3 +4,4 @@
 
 #import <Bugsnag/Bugsnag.h>
 #import "Scenario.h"
+#import "AttachCustomStacktraceHook.h"